backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Document Processing
pypdf2>=3.0.0
python-docx>=1.1.0
markdown>=3.5.0

# Vector DB & Embeddings
faiss-cpu>=1.9.0
numpy>=1.26.0
openai>=1.12.0

# Utilities
tiktoken>=0.5.0
tenacity>=8.2.0
aiofiles>=23.2.0
